# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A banking web app api for banka",
    "version": "1.0.0",
    "title": "Banka App"
  },
  "host": "bankaapp-api.herokuapp.com",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/v1/transactions/{transactionid}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "User can view transactionid",
        "description": "this endpoint allows allows users transactionid",
        "operationId": "transactionid",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transactionid",
            "in": "path",
            "description": "transaction id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}